libmstpm: Add OpenSSL
Add OpenSSL as a git submodule and build it.

As of now, the upstream Microsoft TPM 2.0 requires OpenSSL 1.1.1d or higher.
OpenSSL 3.x support is still under review upstream in a pull request.

Similar to OVMF, we also build OpenSSL 1.1.1q by providing an OpenSSL config
and disabling crypto modules that are not needed.

Later we can move to OpenSSL 3.x.

Signed-off-by: Claudio Carvalho <cclaudio@linux.ibm.com>