libmstpm: Add ms-tpm-2.0-ref
Add the Microsoft TPM as a git submodule and build it.

The SVSM does not need the Microsoft TPM simulator, so we build only the
libplatform.a and libtpm.a.

For now we are disabling the TPM Self Test command (-DSELF_TEST=NO) as it is
failing. This issue is explained and fixed in a separate patch.

Signed-off-by: Claudio Carvalho <cclaudio@linux.ibm.com>