kernel/mm/validate: Introudce a new AtomicValidBitmap
The AtomicValidBitmap takes Vec<PageRef> to store the bitmap, and use
the AtomicU64 to access which allows to modify a bitmap with immutable
reference. This is useful for the case that modify bitmap with just
holding a read lock when using RWLock.

AtomicValidBitmap provides the functions to modify the bitmap or check
the bitmap valid or not for a physical address range.

Co-developed-by: Peter Fang <peter.fang@intel.com>
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>