kernel/tdx/percpu: Introduce Vcpu and TdpVp
Vcpu object to represent TDP guest vcpu and is wrapped in TdpVp. As a TD
can maximum support 3 TDP guests, so each TdPerCpu will have maximum 3
TdpVp objects, and each is represented by using VirtAddr to reduce the
size of TdPerCpu, so that to reduce the memory requirement during
allcoating PerCpu in both stage2 and svsm.

The Vcpu object is initialized in run_tdpvp interface which is used to
run a vcpu. Currently there is no caller for run_tdpvp. The caller will
be added after all the hypervisor code is ready.

Co-developed-by: Jason Chen CJ <jason.cj.chen@intel.com>
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>