kernel/tdx/ioreq: Emulate MMIO request
The MMIO reading/writing instruction is decoded and emulated through
InstrEmul. The MMIO read is emulated by reading the data from virtual
device and then writing back to the corresponding vcpu register or guest
memory. And the MMIO write is emulated by reading the data from the
corresponding vcpu register or guest memory and then writing to the
virtual device.

Currently svsm only emulates vlapic and vioapic, so accessing either
vlapic or vioapic will be handled inside the svsm. For the other virtual
devices, will forward to host through tdvmcall.

As an MMIO instruction might be repeatable, the RIP will be advanced
only when the emulation for this repeat instruction is completed.

Co-developed-by: Vijay Dhanraj <vijay.dhanraj@intel.com>
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>