kernel/tdx/vmexit: Handle tdvmcall MapGpa request
The virtio device models are emulated by host. An elightened TDP guest
can use tdvmcall MapGpa to share its virtual DMA memory with host so
that host can directly emulate virtio device DMA with this shared
memory to accelerate virtio device performance and simplify the svsm.
The shared memory can also be un-shared by this tdvmcall when it is
un-used to guest.

So the tdvmcall MapGpa is handled in svsm by converting the requested
guest memory from private to share, or vice versa.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>