kernel/tdx/tdp: Introduce Tdp to represent a Tdp guest
A TD can maximum support 3 TDP guests, and each TDP guest can be
represented by a Tdp object which contains a TdpVmId and is protected by
a RWLock. So totally there are 3 Tdp objects in the global TDPS slice.

TDPS is initialized by function init_tdps according to the actual number
of TDP guests which is configured during launching TD. The actual number
is between 1 and 3 if TDP is supported, otherwise it should be 0. For
the non-supported Tdp guest, its corresponding init method will not be
called to do the initialization.

The init_tdps currently is still a dead function and will be used by the
up-coming patch.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>