kernel/mm/guestmem: Add GuestMemMap to map guest memory
Guest memory is not mapped in svsm's page table. To let svsm being able
to access guest memory, introduce GuestMemMap structure. GuestMemMap
uses the ptguard to map the desired guest memory in svsm's page table
with a temporary VirtAddr, and convert this VirtAddr to GuestPtr for the
reading and writing.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>