kernel/tdx/vmcs: Introduce Vmcs field for Vcpu
Vmcs is added to Vcpu, which is currently has basic read/write
operations implemented to access L2 VMCS. The read/write operations are
defined according to the data length. vmcs_lib is defined to contain the
basic VMCS field definations and the low level vmcs read/write operations
for Vmcs.

Co-developed-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Signed-off-by: Jason Chen CJ <jason.cj.chen@intel.com>