kernel/tdx/gctx: Add GuestCpuContext to Vcpu
GuestCpuContext represents the vcpu registers. Initially GuestCpuContext
structure contains a TDP context with the format defined in TDP spec. It
provides TDP guest's gpregs, rip, rflags and interrupt_status to TDX
module. Its physicall address is required to be 256Bytes aligned, and
will be used as the input parameter of tdcall TDG.VP.ENTER.

The GuestCpuContext resets the rflags and rip to real-mode value and the
other registers are cleared to be zero. More vcpu registers will be
added in the up-coming patches.

Co-developed-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Signed-off-by: Jason Chen CJ <jason.cj.chen@intel.com>