kernel/mm/validate: Extend ValidBitmap to use PageRef
ValidBitmap currently is used as bitmap for SVSM memory. When it is used
for guest memory, then the bitmap memory size may be larger than the
maximum memory size allocate_pages() can support. To resolve this, use
PageRef as an altertative to store the bitmap. So ValidBitmap can
support using either *mut u64 or Vec<PageRef>. With this change, the
stage2 still init VALID_BITMAP with *mut u64, and svsm will migrate the
bitmap contents from *mut u64 to Vec<PageRef>.

Co-developed-by: Peter Fang <peter.fang@intel.com>
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>