kernel/mm/guestmem: Add helper functions for shared GPA
The guest physical address uses the same bits with svsm physical address
as the encryption bit mask. To check if a GPA is shared or not, check
the encryption bit mask according to the cpu types.

The other helper function is to strip the encryption bit from a GPA
which is similar as strip_c_bit for the PhysAddr.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>