kernel/tdx/tdp: Register VcpuCommBlock to TDPCBS
Introduce a per-VM TDPCBS that maintains an Arc pointer to the BTreeMap,
which contains the Arc pointers for all VcpuCommBlocks. Each Arc pointer
to the BTreeMap is encapsulated by OnceCell, permitting a single
initialization.

Every vcpu registers its VcpuCommBlock with the Tdp's BTreeMap. Once all
vcpus have finished their registration, an Arc pointer to this BTreeMap
is stored in the TDPCBS

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>