kernel/tdx/gmem: Add GuestMemBitmaps to represent accepted and shared pages
Create GuestMemBitmap to manage accepted and shared guest pages, which
is wrapped by a RWLock.

The accepted bitmap is implemented by using AtomicValidBitmap as it can
be set without mutable. This is because accepting a guest page which is
already accepted by another won't cause issue. And accepting won't
change the shared bitmap but only check the shared bitmap. So accepting
a guest page can just take read lock. The pub function accept_guest_mem
can be used to accept a guest memory region.

The shared bitmap is implemented by using ValidBitmap as it can only be
modified with mutable. This is because sharing and accepting should run
exclusive. Share pages will unset the accepted bitmap and accepting
pages which are shared is not allowed. The shared page will be accepted
only when they are coverted back to private. So converting a guest page
should take the write lock. The pub function convert_guest_mem can be
used to convert a guest memory region between private and shared.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>