kernel/tdx/utils: Add helper function to convert svsm page
The function td_convert_kernel_pages can take a range of svsm memory and
convert this memory from private to shared, or reversal by using the
tdvmcall MapGPA. And the corresponding page table mappings should also
be remapped as shared or encrypted. For the direction of from shared to
private, the memory should be accepted before it can be accessed by
svsm.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>