kernel/tdx/virq: Add support to inject exception
When svsm needs to inject an exception to TDP guest, the exception
vector and error code is queued in Virq according to double faults and
triple faults conditions and then add a pending vcpu request to
VcpuReqFlags. The queued exception is injected during handle the pending
vcpu request before vmenter. The exception has higher priority than NMI
so is injected first.

Co-developed-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Signed-off-by: Jason Chen CJ <jason.cj.chen@intel.com>