kernel/tdx/instr_emul: Add instruction emulation support
Provide method emulate_instr to provide IO instruction emulation support
for both MMIO and PIO. Method takes IoOperation as input, and emulate
differently according to the OpType. Basically the read IoOperation is
emulated by writing the read data in the IoOperation to the
corresponding registers or guest memory. And the write IoOperation is
emulated by reading the data to be written from the corresponding
registers or guest memory to IoOperation.

For the repeat instruction, set retry flag to indicate TDP guest to
re-execute this instruction so that it can be emulated again until the
repeat count reaches to 0.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>