kernel/tdx/vmexit: Handle cpuid vmexit
The Vcpuid is a per TDP guest object, which is one of the element in the
Tdp structure. For the global cpuid leaves, the emulated result is
directly from the BTreeMap entries of the Vcpuid as they are common
between vcpus. For the per-cpu cpuid leaves, emulate differently
according the leaves. Currently leaf 0x01, 0x0b 0x0d, 0x19 and
0x80000001 are emulated according to corresponding features. The other
per-cpu leaves are emulated by returning svsm's cpuid.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>