kernel/tdx/vlapic: Support injecting interrupt to remote vcpu
Introduce VirqBitmap in VcpuCommBlock to support injecting interrupt to
a remote vcpu. The source vcpu can make a pending virq with vector and
trigger level, which will be set into the VirqBitmap of the dest vcpu's
VcpuCommBlock, and kick the dest vcpu to handle.

The VirqBitmap will be scanned by the dest vcpu to sync all the pending
vector/level information to its vlapic before handling interrupt event
injection, so that can be injected to TDP guest in the subsequent code
path.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>