kernel/mm/pagetable: Disable page splitting
Due to PerCpu page table sharing and the shared bit, splitting a 2M page
becomes tricky in SMP cases. Map the SVSM heap and guest memory regions
in 4K-only mode so that the shared bits across PTEntry's can be toggled
simultaneously.

Don't allow 2M pages to be split for now.

Signed-off-by: Peter Fang <peter.fang@intel.com>